progetto/WebApp/requirements.txt
Flask>=3.0,<4.0
Flask-Session>=0.6,<1.0
mysql-connector-python>=8.0,<9.0
